Extract build into module to make it easier to write a webpack loader.
